+++ spectrochempy/analysis/_base.py
@@ -655,6 +655,7 @@ class DecompositionAnalysis(AnalysisConfigurable):
     See Also
     --------
     EFA : Perform an Evolving Factor Analysis (forward and reverse).
+    FastICA : Perform Independent Component Analysis with a fast algorith.
     IRIS : Integral inversion solver for spectroscopic data.
     MCRALS : Perform MCR-ALS of a dataset knowing the initial :math:`C` or :math:`S^T` matrix.
     NMF : Non-Negative Matrix Factorization.